Change in database shows up in app & Change in app shows up in database [Issue 1]
