+++ src/App.js
@@ -1,21 +1,17 @@
 import React, { useState } from 'react';
-import Items from './Items';
 import ItemsInput from './ItemsInput';
+import Listener from './Listener';
 import './App.css';
 // import logo from './logo.svg';
 
 function App() {
+  //To display the items from database
   const [itemsDisplay, setItemsDisplay] = useState(false);
-  // const [itemName, setItemName] = useState();
 
   return (
     <div className="App">
-      {/* <button onClick={() => setItemsDisplay(true)}>Click to get items!</button>
-    {itemsDisplay && <Items/>} */}
-
       <button onClick={() => setItemsDisplay(true)}>Click to get items!</button>
-      {itemsDisplay && <Items />}
-
+      {itemsDisplay && <Listener />}
       <ItemsInput />
 
       {/* <header className="App-header">
